refactor: move package to de.epischel.javamember
