+++ app/build.gradle
@@ -34,7 +34,7 @@ java {
 
 application {
     // Define the main class for the application.
-    mainClass = 'org.example.App'
+    mainClass = 'de.epischel.javamember.App'
 }
 
 tasks.named('test') {
