+++ app/src/main/java/de/epischel/javamember/App.java
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example;
+package de.epischel.javamember;
 
 public class App {
     public String getGreeting() {
